refactor: move header to core
