+++ YouTube-client-app/src/app/app.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 
-import { HeaderComponent } from './components/header/header.component';
+import { HeaderComponent } from './core/components/header/header.component';
 import { MainPageComponent } from './youtube/main-page/main-page.component';
 
 @Component({
