+++ YouTube-client-app/src/app/core/components/header/header.component.html
