+++ YouTube-client-app/src/app/core/components/header/header.component.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { Component, EventEmitter, Output } from '@angular/core';
 
-import { ButtonComponent } from '../../shared/button/button.component';
+import { ButtonComponent } from '../../../shared/button/button.component';
 import { FilterComponent } from './filters/filter.component';
 import { LoginInfoComponent } from './login-info/login-info.component';
 import { SearchInputComponent } from './search-input/search-input.component';
