+++ YouTube-client-app/src/app/core/components/header/login-info/login-info.component.scss
@@ -7,7 +7,7 @@
   &__img {
     width: 100%;
     height: 100%;
-    background-image: url("./../../../../assets/img/login.png");
+    background-image: url("./../../../../../assets/img/login.png");
     background-repeat: no-repeat;
     background-position: center;
     background-size: 100%;
