+++ YouTube-client-app/src/app/core/components/header/login-info/login-info.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { ButtonComponent } from '../../../shared/button/button.component';
+import { ButtonComponent } from '../../../../shared/button/button.component';
 
 @Component({
   selector: 'app-login-info',
