+++ YouTube-client-app/src/app/core/components/header/pipes/search-pipe.pipe.spec.ts
