+++ YouTube-client-app/src/app/core/components/header/search-input/search-input.component.scss
