+++ YouTube-client-app/src/app/core/components/header/search-input/search-input.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 
-import { ButtonComponent } from '../../../shared/button/button.component';
-import { CardShowService } from '../../../youtube/main-page/cards-list/services/card-show-service.service';
+import { ButtonComponent } from '../../../../shared/button/button.component';
+import { CardShowService } from '../../../../youtube/main-page/cards-list/services/card-show-service.service';
 import { SearchStateService } from '../services/search-state.service';
 
 @Component({
