+++ YouTube-client-app/src/app/core/components/header/settings-button/setting-img.component.ts
