+++ YouTube-client-app/src/app/core/components/header/settings-button/settings-button.component.scss
