+++ YouTube-client-app/src/app/core/components/header/settings-button/settings-button.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { ButtonComponent } from '../../../shared/button/button.component';
+import { ButtonComponent } from '../../../../shared/button/button.component';
 import { SettingImgComponent } from './setting-img.component';
 
 @Component({
