+++ YouTube-client-app/src/app/youtube/main-page/cards-list/cards-list.component.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 
-import { SearchPipePipe } from '../../../components/header/pipes/search-pipe.pipe';
-import { SearchStateService } from '../../../components/header/services/search-state.service';
+import { SearchPipePipe } from '../../../core/components/header/pipes/search-pipe.pipe';
+import { SearchStateService } from '../../../core/components/header/services/search-state.service';
 import { CardItemComponent } from './card-item/card-item.component';
 import { CardItemModel } from './models/card-item.model';
 import { CardDataService } from './services/cardData-service.service';
